app/requirements.txt
paho-mqtt
psycopg[binary]
httpx